+++ tutorial/03.Juicer_Hi-C/QuickStart_juicer.sh
@@ -6,20 +6,21 @@ gene=refFlat.$build.txt
 bwaindex=bwa-indexes/$build
 ncore=64
 
-sing="apptainer exec --nv --bind /work,/work2,/work3 /work/SingularityImages/custardpy.3.2.0.sif"
+sing="apptainer exec --nv --bind /work,/work2,/work3 /work/SingularityImages/custardpy.3.4.1.sif"
 
 fastq_post="_"  # "_" or "_R"
 enzyme=MboI
 
-for cell in Control siCTCF siRad21 siNIPBL
+for cell in Control #siCTCF siRad21 siNIPBL
 do
     fqdir=fastq/$cell
-#    $sing custardpy_juicer -p $ncore -a $gene -b $build -g $gt -i $bwaindex -e $enzyme -z $fastq_post $fqdir $cell
+    $sing custardpy_juicer -p $ncore -a $gene -b $build -g $gt -i $bwaindex -e $enzyme -z $fastq_post $fqdir $cell
     odir=CustardPyResults/Juicer_$build/$cell
     hic=$odir/aligned/inter_30.hic
     norm=SCALE
     $sing custardpy_process_hic -p $ncore -n $norm -g $gt -a $gene $hic $odir
 done
+
 exit
 odir=CustardPyResults/Juicer_$build/
 norm=SCALE